Disable TPL and run the pipeline.
The TPL images are producing 404s because the TPL site has changed. The crawler will have to be updated. Until it is: disable TPL.

(Also note that both the pipeline code and docs are out of date. For example, the code looks for files that no other code produces, and the docs mention files that no code produces. Still, this run produced a useful result!)